third_party/blink/web_tests/fast/canvas/OffscreenCanvas-2d-placeholder-willReadFrequently.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

// Testing that willReadFrequently works the same with HTMLCanvasElement and
// OffscreenCanvas with a placeholder canvas
function do_test(t, context_options) {
  const canvas = document.createElement('canvas');
  const offscreen_canvas = canvas.transferControlToOffscreen();
  const ctx = offscreen_canvas.getContext('2d', context_options);
  const reference_canvas = document.createElement('canvas');
  const reference_ctx = reference_canvas.getContext('2d', context_options);
  const img = document.createElement('img');
  img.src = "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAkAAAAJCAYAAADgk" +
  "QYQAAAAGklEQVQYlWNgYGD4j4z/////H12MYVQRUYoAkYZrlWt0UekAAAAASUVORK5CYII=";
  img.onload = () => {
    ctx.drawImage(img, 0, 0, 300, 300);
    reference_ctx.drawImage(img, 0, 0, 300, 300);
    const image_data = ctx.getImageData(0, 0, 300, 300);
    const expected_image_data = reference_ctx.getImageData(0, 0, 300, 300);
    t.step(() => {
      assert_array_equals(image_data.data, expected_image_data.data);
    });
    t.done();
  };
}

async_test(t => {
  do_test(t, { willReadFrequently: true });
}, "willReadFrequently true");

async_test(t => {
  do_test(t, { willReadFrequently: false });
}, "willReadFrequently false");

async_test(t => {
  do_test(t, {});
}, "willReadFrequently default");

</script>
